+++ src/app/shared/services/annotation-builder.service.ts
@@ -1,8 +1,9 @@
 import { Injectable, inject } from '@angular/core';
-import { Vector3 } from 'three';
+import { Box3, Vector3 } from 'three';
 import { AnnotationType } from '../models/annotation-type.enum';
 import { BaseAnnotation } from '../models/annotations/base.annotation';
 import { CeilingHeightAnnotation } from '../models/annotations/ceiling-height.annotation';
+import { CrossSectionAnnotation, vectorAngleAroundY } from '../models/annotations/cross-section.annotation';
 import { MeasureDistanceAnnotation } from '../models/annotations/measure-distance.annotation';
 import { IMetadataBaseAnnotationV0 } from '../models/manifest/types.v0';
 import { LocalizeService } from './localize.service';
@@ -59,6 +60,57 @@ export class AnnotationBuilderService {
       identifier,
       floorPointLocal,
       additionalPoints,
+    );
+  }
+
+  buildCrossSection(
+    identifier: string,
+    dimensions: Vector3,
+    centerPoint: Vector3,
+    angleToNorthAroundY: number,
+  ) {
+    return new CrossSectionAnnotation(
+      identifier,
+      dimensions,
+      centerPoint,
+      angleToNorthAroundY,
+    );
+  }
+
+  buildCrossSectionFromCrosslineAndBoundingBox(
+    identifier: string,
+    origin: Vector3,
+    dest: Vector3,
+    depth: number,
+    boundingBoxOfModels: Box3
+  ) {
+    const { min: { y: minY }, max: { y: maxY } } = boundingBoxOfModels;
+
+    const originAtMin = origin.clone().setY(minY);
+    const destAtMin = dest.clone().setY(minY);
+
+    const height = maxY - minY;
+    const width = originAtMin.distanceTo(destAtMin);
+    const dimensions = new Vector3(width, height, depth);
+
+    const vectorFromOriginToDest =
+      destAtMin
+        .clone()
+        .sub(originAtMin);
+
+    const angleToNorthOfBoxNormal =
+      vectorAngleAroundY(new Vector3(1, 0, 0), vectorFromOriginToDest);
+
+    const centerPoint = originAtMin
+      .clone()
+      .add(vectorFromOriginToDest.clone().divideScalar(2))
+      .add(new Vector3(0, height / 2, 0));
+
+    return new CrossSectionAnnotation(
+      identifier,
+      dimensions,
+      centerPoint,
+      angleToNorthOfBoxNormal,
     );
   }
 }
